refactor(test-suite): trigger fhe materials via docker container
