+++ test-suite/fhevm/docker-compose/host-node-docker-compose.yml
@@ -1,6 +1,6 @@
 services:
   host-node:
-    container_name: fhevm-host-node
+    container_name: host-node
     image: ghcr.io/foundry-rs/foundry:v1.3.5
     env_file:
       - ../env/staging/.env.host-node.local
